requirements.txt
django==3.2.4
google-api-python-client==2.33.0
google-auth-oauthlib==0.4.6
google-auth-httplib2==0.1.0
django-picklefield==3.0.1
django-model-utils==4.2.0
django-crispy-forms==1.13.0
crispy-bootstrap5==0.6.0
celery==4.4.2
django-db-logger==0.1.10
python-dotenv==0.14.0
redis==3.5.3
django-celery-beat==2.2.0
gunicorn==20.0.4
mysqlclient==2.0.1
django-mysql==3.8.1
python-gettext==4.0.0